+++ CHANGELOG.md
@@ -1,9 +1,26 @@
 # Change Log
+### 0.0.1
+* Initial release with basic features
+### 0.0.2
+* Included a wonderful icon
+### 0.0.3
 
+* added other views to see individual tested master masked and isolated
+* adedd hover to easy toggle on tested and master image
+* change icon on primary sidebar
+### 1.0.0
+* fix size aspect ratio
+* preparing code to new features
+### 1.0.1
+* added clear function
+* tree view by project
+### 1.0.2
+* projects and itens are sorted
+* fix duplicated itens on tree
+### 1.0.3
+* automatic refresh when detect new failures
 All notable changes to the "golden-diff" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
 
-- Initial release